requirements.txt
google-cloud-aiplatform
google-cloud-logging
streamlit